### 2025-04-03 #7926 @arfon
@ThijsVroegh – could you take a look a [this list of potential reviewers](https://reviewers.joss.theoj.org/lookup) and identify a few people who would be good candidates to review this submission?

Also, I note there are a few issues with your paper and citation formatting. Please review these guidelines to see how to get them rendering properly: https://pandoc.org/MANUAL.html#extension-citations
