+++ paper/paper.bib
@@ -20,7 +20,7 @@
 
 @book{bambach1999,
   title={Drawing and painting in the Italian Renaissance workshop: theory and practice, 1300-1600},
-  author={Bambach, Carmen},  
+  author={Bambach, Carmen},
   year={1999},
   publisher={Cambridge University Press}
 }
@@ -47,7 +47,8 @@
   number={11},
   pages={2385--2401},
   year={2009},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/tip.2009.2025923}}
 }
 
 @article{smith2024,
@@ -58,7 +59,8 @@
   number={6},
   pages={e0302962},
   year={2024},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pone.0302962}}
 }
 
 @book{stork2023,
@@ -75,7 +77,8 @@
   number={4},
   pages={48--59},
   year={2024},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.visinf.2024.11.001}}
 }
 
 @article{ugail2023,
@@ -86,7 +89,8 @@
   number={1},
   pages={1--15},
   year={2023},
-  publisher={SpringerOpen}
+  publisher={SpringerOpen},
+  doi={10.1186/s40494-023-01094-0}}
 }
 
 @ARTICLE{wang2004,
