### 2025-04-24 #7926 @ThijsVroegh
Hi @arfon, I updated the paper with the suggested DOIs, and looked for potential reviewers at https://reviewers.joss.theoj.org/reviewers. I used the search terms:
- python
- image processing, computer vision

From the small list of people, I made a shortlist of potential reviewers:

- Name	[Github]
- Animikh Aich      [animikhaich]
- Amit Soni Arya	[amitsoniR]
- Atif-Anwer	        [Atif-Anwer]
- mgezer	                [mgezer]
- Tobias Uelwer	 [tuelwer]
- mohammadul	  [mohammadul]

I hope this will work!
