### 2025-05-12 #7926 @arfon
@mgezer @mohammadul @tuelwer :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _UnRaphael: A Python pipeline tool for analysing the similarity between paintings and their reproductions_

Many thanks
Arfon
