### 2025-05-12 4d7b264
Add commands to run Docker image
