### 2025-05-12 bf8ae4a
Add commands to run dashboard
