+++ src/unraphael/dash/image_clustering.py
@@ -52,6 +52,7 @@ MSE_NUMERATOR = 1000.0
 NUM_THREADS = 8
 
 import os
+
 torch.classes.__path__ = []  # Simple fix
 # Alternatively: torch.classes.__path__ = [os.path.join(torch.__path__[0], torch.classes.__file__)]
 
@@ -546,45 +547,46 @@ def calculate_cw_ssim_similarity(i1: np.ndarray, i2: np.ndarray) -> float:
 
     Strong for handling small geometric distortions in structural
     comparison.
-    
+
     Implementation follows the original CW-SSIM algorithm from Wang and Simoncelli.
     """
-        
+
     # Convert images to grayscale if needed
     if len(i1.shape) == 3:
         i1_gray = np.mean(i1, axis=2).astype(np.float32)
     else:
         i1_gray = i1.astype(np.float32)
-    
+
     if len(i2.shape) == 3:
         i2_gray = np.mean(i2, axis=2).astype(np.float32)
     else:
         i2_gray = i2.astype(np.float32)
-    
+
     # Flatten the arrays for wavelet transform
     sig1 = i1_gray.flatten()
     sig2 = i2_gray.flatten()
-    
+
     # Define custom ricker wavelet function as fallback
     def custom_ricker(points, a):
         """
         Return a Ricker wavelet (Mexican hat wavelet) of length 'points' with parameter 'a'.
-        
-        This is a custom implementation that can be used when scipy.signal.ricker 
+
+        This is a custom implementation that can be used when scipy.signal.ricker
         or scipy.signal.windows.ricker are not available.
         """
         A = 2 / (np.sqrt(3 * a) * np.pi**0.25)
         wsq = a**2
         vec = np.arange(0, points) - (points - 1.0) / 2
         xsq = vec**2
-        mod = (1.0 - xsq / wsq)
+        mod = 1.0 - xsq / wsq
         gauss = np.exp(-xsq / (2 * wsq))
         return A * mod * gauss
-    
+
     # Use the correct wavelet function (handle both old and new SciPy versions)
     try:
         # Try new SciPy version (windows module)
         from scipy.signal.windows import ricker
+
         wavelet = ricker
     except ImportError:
         try:
@@ -593,32 +595,32 @@ def calculate_cw_ssim_similarity(i1: np.ndarray, i2: np.ndarray) -> float:
         except AttributeError:
             # Fall back to custom implementation if neither is available
             wavelet = custom_ricker
-    
+
     # Set width parameter for the wavelet transform
     widths = np.arange(1, 30)
-    
+
     # Perform the continuous wavelet transform
     cwtmatr1 = signal.cwt(sig1, wavelet, widths)
     cwtmatr2 = signal.cwt(sig2, wavelet, widths)
-    
+
     # Small constant for stability
     k = 0.01
-    
+
     # Compute the first term (magnitude)
     c1c2 = np.multiply(abs(cwtmatr1), abs(cwtmatr2))
     c1_2 = np.square(abs(cwtmatr1))
     c2_2 = np.square(abs(cwtmatr2))
     num_ssim_1 = 2 * np.sum(c1c2, axis=0) + k
     den_ssim_1 = np.sum(c1_2, axis=0) + np.sum(c2_2, axis=0) + k
-    
+
     # Compute the second term (phase)
     c1c2_conj = np.multiply(cwtmatr1, np.conjugate(cwtmatr2))
     num_ssim_2 = 2 * np.abs(np.sum(c1c2_conj, axis=0)) + k
     den_ssim_2 = 2 * np.sum(np.abs(c1c2_conj), axis=0) + k
-    
+
     # Construct the result
[…]
