+++ Dockerfile
@@ -2,7 +2,7 @@
 # docker build -t ghcr.io/decodingraphael/unraphael .
 # Push to GitHub Container Registry:
 # docker push ghcr.io/decodingraphael/unraphael
-# Pull:
+# Run:
 # docker run -p 8501:8501 ghcr.io/decodingraphael/unraphael
 FROM python:3.12-slim
 
