### 2025-05-12 3569292
pre-commit fix
With `uvx pre-commit run --all-files`
