### 2025-05-12 8a41f70
Set max line length to 270 + format
This makes `uvx pre-commit run --all-files` happy
