+++ src/unraphael/dash/image_clustering.py
@@ -101,9 +101,7 @@ def compute_mean_sharpness(images: list[np.ndarray]) -> float:
     return mean_sharpness / len(images)
 
 
-def normalize_brightness_set(
-    images: list[np.ndarray], mean_brightness: float
-) -> list[np.ndarray]:
+def normalize_brightness_set(images: list[np.ndarray], mean_brightness: float) -> list[np.ndarray]:
     """Normalize brightness of all images in the set to the mean brightness."""
     normalized_images = []
     for img in images:
@@ -112,11 +110,7 @@ def normalize_brightness_set(
             l_channel, a_channel, b_channel = cv2.split(img_lab)
             current_brightness = np.mean(l_channel)
             print(f'Original brightness: {current_brightness}')
-            l_channel = (
-                (l_channel * (mean_brightness / current_brightness))
-                .clip(0, 255)
-                .astype(np.uint8)
-            )
+            l_channel = (l_channel * (mean_brightness / current_brightness)).clip(0, 255).astype(np.uint8)
             normalized_img_lab = cv2.merge([l_channel, a_channel, b_channel])
             normalized_img = cv2.cvtColor(normalized_img_lab, cv2.COLOR_LAB2BGR)
             print(f'Normalized brightness: {np.mean(l_channel)}')
@@ -125,11 +119,7 @@ def normalize_brightness_set(
             l_channel = img
             current_brightness = np.mean(l_channel)
             print(f'Original brightness: {current_brightness}')
-            normalized_img = (
-                (l_channel * (mean_brightness / current_brightness))
-                .clip(0, 255)
-                .astype(np.uint8)
-            )
+            normalized_img = (l_channel * (mean_brightness / current_brightness)).clip(0, 255).astype(np.uint8)
             print(f'Normalized brightness: {np.mean(normalized_img)}')
 
         normalized_images.append(normalized_img)
@@ -146,9 +136,7 @@ def normalize_contrast_set(images: list[np.ndarray], mean_contrast: float) -> li
             l_channel, a_channel, b_channel = cv2.split(img_lab)
             current_contrast = np.std(l_channel)
             print(f'Original contrast: {current_contrast}')
-            l_channel = (
-                (l_channel * (mean_contrast / current_contrast)).clip(0, 255).astype(np.uint8)
-            )
+            l_channel = (l_channel * (mean_contrast / current_contrast)).clip(0, 255).astype(np.uint8)
             normalized_img_lab = cv2.merge([l_channel, a_channel, b_channel])
             normalized_img = cv2.cvtColor(normalized_img_lab, cv2.COLOR_LAB2BGR)
             print(f'Normalized contrast: {np.std(l_channel)}')
@@ -157,9 +145,7 @@ def normalize_contrast_set(images: list[np.ndarray], mean_contrast: float) -> li
             l_channel = img
             current_contrast = np.std(l_channel)
             print(f'Original contrast: {current_contrast}')
-            normalized_img = (
-                (l_channel * (mean_contrast / current_contrast)).clip(0, 255).astype(np.uint8)
-            )
+            normalized_img = (l_channel * (mean_contrast / current_contrast)).clip(0, 255).astype(np.uint8)
             print(f'Normalized contrast: {np.std(normalized_img)}')
 
         normalized_images.append(normalized_img)
@@ -167,9 +153,7 @@ def normalize_contrast_set(images: list[np.ndarray], mean_contrast: float) -> li
     return normalized_images
 
 
[…]
