+++ pyproject.toml
@@ -118,7 +118,7 @@ module = ["yaml.*", "toml.*"]
 ignore_missing_imports = true
 
 [tool.ruff]
-line-length = 96
+line-length = 270
 target-version = "py310"
 extend-include = ["*.ipynb"]
 

[…]
