### 2025-06-15 #7926 @arfon
@Atif-Anwer @amitsoniR @animikhaich @JasonGellis 👋 would any of you be willing to review this submission for JOSS? The submission under consideration is _UnRaphael: A Python pipeline tool for analysing the similarity between paintings and their reproductions_

Many thanks
Arfon
