### 2025-07-15 #7926 @Atif-Anwer
Hi @arfon! Sure, I'll be able to review the paper around the next week onwards (21st July +).

Thanks and with best regards,
Atif
