### 2025-08-31 #8900 @editorialbot
[REVIEW]: UnRaphael: A Python pipeline tool for analysing the similarity between paintings and their reproductions
**Submitting author:** @ThijsVroegh (<a href="http://orcid.org/0000-0002-6311-0798">T. Vroegh</a>)
**Repository:** https://github.com/DecodingRaphael/unraphael
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3
