### 2025-09-14 #8900 @Atif-Anwer
@arfon  My review for the paper is attached below:

The submission is a Python library titled "Unraphael" which provides a python pipeline for analysing similarity between Raphael paintings and their reproductions, and is primarily aimed at art historical research. The pipeline implemented by a Streamlit power GUI provides several tools and methodologies for pre-processing and postprocessing and analyzing images.
Overall, the package is well documented and available at https://unraphael.readthedocs.io/en/latest/, and provides a clear and simple way to install and use the package using pip. The Examples provided to test each page are quite helpful for a new user and non-specialists to understand and use the tools provided.

However, there are several things that need to be addressed before I can recommend JOSS to accept this submission:

## Major issues
- Currently, the paper mentions several procedures included as part of the tool including Pre-processing, Image Similarity, Object Segemenation+Detection, Alignment and Clustering pages. Running `unraphael-dash` in the terminal works fine, but does not show/run the Alignment, Clustering or Ratio pages in the streamlit dashboard.

- The pip package is an older release version 0.21 (dated Jun 13, 2024) where as the repo package version is 0.3 (dated 6 Feb 2025). Since the GitHub repo requires the unraphael package from pip to function, there are several errors and the code from the GitHub repo is not able to run, and I was unable to test it. Comparing the two version shows several new imports and function added in v0.3, which are missing in the installed pip version of unraphael.

- Image Similarity page requires Loading a config yaml file, but it is unclear and unintuitive as to what config is required. Also, not everyone may be familiar with yaml format and its formatting etc. It may be useful to add the configuration as Sliders/Dropdown on the top of the Image Similarity Page so that the settings can be varied by the user easily. The config can be generated and used automatically on execution of the similarity procedures.

- Similarly, the Feature Extraction and Image Matching have editable textboxes for config values; and even accept random text (that just crashes the application). This can lead to inadvertent crashes and confusion for users and is recommended to be changed to something better.

- The app uses Ultralytics YOLO for detection, segmentation, and pose estimation; but the lines 36-41 in the paper mention avoiding these 'black-box' methods which is contradictory and incorrect. This needs to be rectified and avoided.

- In the ratio's page, there is an xlsx file required to be uploaded containing painting dimensions, but there is no example or detail provided on what it should contain in what format? Not only can these dimensions can be inferred directly from the image, but It may be easier to provide table entries in the application page directly instead of using an excel file. Or provide a sample with the required rows/columns for the users to fill.

- The provided streamlit website (https://unraphael.streamlit.app) does not function.

## Minor suggestions for QOL improvement
- The main github repo Requirements.txt is quite sparse for a clean installation. It should include all packages being used for a user to clean install in a new virtual environment that are required to run the code locally. The dash package requirements.txt should be copied to the main maybe?

- The figure as part of the paper needs to be updated to show the latest version of the app (increment/decrement boxes instead of sliders)

- The Sidebar Menus and other acronyms (SIFT, ORB etc) need proper capitalization.

- There are several warning probably due to outdated package versions (?) that needs to be removed.

- If no config yaml is loaded but images are loaded, the app crashes. This is also not intuitive to figure out, and should not be allowed.

- The Image limit is set to 200MB however loading a 95MB image crashes the app with error `PIL.Image.DecompressionBombError: Image size (301404458 pixels) exceeds limit of 178956970 pixels, could be decompression bomb DOS attack.` despite having enough RAM/processing compute available. Image limits need to be checked for a more robust error check.

- The detection model on the Example reports the `TEDDY BEAR` class, but works on other Raphael images.

- In the pre-processes page, The Extracted image is not the one that is pre-processed (it is segmented from the original uploaded image instead)? Is this correct/desired or by accident?

- Other than yolo for keypoints, I would also suggest testing mediapipe for keypoints if it helps, as it is easy to integrate and has very decent keypoint detection results from experience.

I would like to request the authors to ensure that the package versions installed by pip are compatible with the updated GitHub repo code. And the features listed should be the same whether the code is run from the pip package (unraphael-dash) or the repo locally. The code is very well documented, and can be really useful for the intended target audience, but only if they are able to run it easily.
